+++ CHANGELOG.md
@@ -119,6 +119,7 @@
 * Fix missing preview field label. Fixes [UILD-541].
 * Hide preview in complex lookups. Fixes [UILD-709].
 * Show a toast after saving a resource. Fixes [UILD-563].
+* Fix the broken search results table. Fixes [UILD-788].
 * Importing an empty file treated as an error. Fixes [UILD-692].
 * Fix CSS rule leaking out of this module. Fixes [UILD-796].
 
@@ -239,6 +240,7 @@
 [UILD-541]:https://folio-org.atlassian.net/browse/UILD-541
 [UILD-709]:https://folio-org.atlassian.net/browse/UILD-709
 [UILD-563]:https://folio-org.atlassian.net/browse/UILD-563
+[UILD-788]:https://folio-org.atlassian.net/browse/UILD-788
 [UILD-692]:https://folio-org.atlassian.net/browse/UILD-692
 [UILD-796]:https://folio-org.atlassian.net/browse/UILD-796